+++ package.json
@@ -14,6 +14,7 @@
     "prop-types": "^15.6.2",
     "react": "^16.8.0",
     "react-dom": "^16.6.0",
+    "react-pose": "^4.0.8",
     "react-scripts": "2.1.0",
     "react-spring": "^8.0.19",
     "socket.io-client": "^2.1.1",
