--- src/components/RoomArea.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-/**
- * React component that shows the current room members and roll history
- * @constructor
- */
-const RoomArea = () => {
-  return <div />;
-};
-
-export default RoomArea;
