+++ src/components/index.tsx
@@ -2,5 +2,6 @@ import DiceButtons from "./DiceButtons";
 import Logo from "./Logo";
 import TopControls from "./TopControls";
 import MainView from "./MainView";
+import TransitionRouter from "./TransitionRouter";
 
-export { DiceButtons, Logo, TopControls, MainView };
+export { DiceButtons, Logo, TopControls, MainView, TransitionRouter };
